% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generate-stan.R
\name{create_hierarchical_profiles}
\alias{create_hierarchical_profiles}
\title{Generate constrained mastery profiles for hierarchical measurement model}
\usage{
create_hierarchical_profiles(attributes, hierarchy, att_labels)
}
\arguments{
\item{attributes}{Positive integer. The number of attributes being measured.}

\item{hierarchy}{A tibble specifying the attribute hierarchy (via
\code{\link[ggdag:tidy_dagitty]{ggdag::tidy_dagitty()}}).}

\item{att_labels}{A tibble mapping the generic attribute names to the
attribute names from the user-specified Q-matrix.}
}
\value{
A \link[tibble:tibble-package]{tibble} with all possible attribute
mastery profiles. Each row is a profile, and each column indicates whether
the attribute in that column was mastered (1) or not mastered (0). Thus,
the tibble will have \code{2^attributes} rows, and \code{attributes} columns.
}
\description{
Given an attribute hierarchy, generate the consistent patterns of attribute
mastery.
}
\examples{
create_hierarchical_profiles(3, ggdag::tidy_dagitty(" dag { x -> y -> z } "),
                             att_labels = tibble::tibble(att = c("att1",
                                                                 "att2",
                                                                 "att3"),
                                                         att_label =c("A1",
                                                                      "A2",
                                                                      "A3")))
create_hierarchical_profiles(4,
                             ggdag::tidy_dagitty(" dag {x -> y -> z -> a} "),
                             att_labels = tibble::tibble(att = c("att1",
                                                                 "att2",
                                                                 "att3",
                                                                 "att4"),
                                                         att_label =c("A1",
                                                                      "A2",
                                                                      "A3",
                                                                      "A4")))
}
