% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmstan-package.R
\docType{package}
\name{dcmstan-package}
\alias{dcmstan}
\alias{dcmstan-package}
\title{dcmstan: Generate 'Stan' Code for Estimating Diagnostic Classification Models}
\description{
What the package does (one paragraph). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dcmstan.r-dcm.org}
  \item \url{https://github.com/r-dcm/dcmstan}
  \item Report bugs at \url{https://github.com/r-dcm/dcmstan/issues}
}

}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})

Authors:
\itemize{
  \item Jeffrey Hoover \email{jeffrey.c.hoover@gmail.com} (\href{https://orcid.org/0000-0002-0276-0308}{ORCID})
  \item Auburn Jimenez \email{auburn.jimenez34@gmail.com} (\href{https://orcid.org/0000-0002-7072-2960}{ORCID})
}

Other contributors:
\itemize{
  \item University of Kansas [copyright holder]
  \item Institute of Education Sciences [funder]
}

}
\keyword{internal}
