% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-class-model-priors.R
\name{default_dcm_priors}
\alias{default_dcm_priors}
\title{Default priors for diagnostic classification models}
\usage{
default_dcm_priors(measurement_model = NULL, structural_model = NULL)
}
\arguments{
\item{measurement_model}{A \link[=measurement-model]{measurement model} object.}

\item{structural_model}{A \link[=structural-model]{structural model} object.}
}
\value{
A \code{dcmprior} object.
}
\description{
View the prior distributions that are applied by default when using a given
measurement and structural model.
}
\examples{
default_dcm_priors(lcdm(), unconstrained())
default_dcm_priors(dina(), independent())
default_dcm_priors(lcdm(), loglinear())
}
