% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-methods-generate-stan.R
\name{generate_stan}
\alias{generate_stan}
\title{Title}
\usage{
generate_stan(x, ...)
}
\arguments{
\item{x}{A model specification (e.g., \code{\link[=dcm_specify]{dcm_specify()}} object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Something
}
