% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-class-model-components.R
\name{measurement-model}
\alias{measurement-model}
\alias{lcdm}
\alias{dina}
\alias{dino}
\alias{crum}
\title{Measurement models for diagnostic classification}
\usage{
lcdm(max_interaction = Inf)

dina()

dino()

crum()
}
\arguments{
\item{max_interaction}{For the LCDM, the highest item-level interaction to
include in the model.}
}
\value{
A measurement model object.
}
\description{
The measurement model defines how the items relate to the attributes. The
currently supported options for measurement models are:
loglinear cognitive diagnostic model (LCDM); deterministic input, noisy "and", gate (DINA); deterministic input, noisy "or" gate (DINO); and compensatory, reparameterized unified model (C-RUM).
See details for additional information on each model.
}
\details{
Additional details...
}
\examples{
lcdm()

lcdm(max_interaction = 2)

dina()
}
