% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-class-model-components.R
\name{measurement}
\alias{measurement}
\alias{model-components}
\alias{structural}
\alias{LCDM}
\alias{DINA}
\alias{DINO}
\alias{CRUM}
\alias{UNCONSTRAINED}
\alias{INDEPENDENT}
\title{Class constructors for model components}
\usage{
measurement(model = "lcdm", model_args = list())

structural(model = "unconstrained", model_args = list())

LCDM(model = "lcdm", model_args = list())

DINA(model = "lcdm", model_args = list())

DINO(model = "lcdm", model_args = list())

CRUM(model = "lcdm", model_args = list())

UNCONSTRAINED(model = "unconstrained", model_args = list())

INDEPENDENT(model = "unconstrained", model_args = list())
}
\arguments{
\item{model}{The type of measurement model to be used. Must be one of
lcdm, dina, dino, or crum.}

\item{model_args}{A named list of arguments to be passed on to the
corresponding \verb{meas_*()} function.}
}
\value{
An \link[S7:new_class]{S7 object constructor}.
}
\description{
The class constructors for model components are exported to facilitate the
defining of methods in other packages. We do not expect or recommend calling
these functions directly. To define and create measurement and structural
models, see \code{\link{measurement-model}} and \code{\link{structural-model}}.
}
