% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-class-model-priors.R
\name{prior}
\alias{prior}
\title{Define a prior distributions}
\usage{
prior(distribution, type, coefficient = NA, lower_bound = NA, upper_bound = NA)
}
\arguments{
\item{distribution}{A distribution statement for the prior
(e.g., \code{normal(0, 2)}). For a complete list of available distributions, see
the \emph{Stan} documentation at \url{https://mc-stan.org/docs/}.}

\item{type}{The type of parameter to apply the prior to. Parameter types will
vary by model. Use \code{\link[=get_parameters]{get_parameters()}} to see list of possible types for
the chosen model.}

\item{coefficient}{Name of a specific parameter within the defined parameter
type. If \code{NA} (the default), the prior is applied to all parameters within
the type.}

\item{lower_bound}{Optional. The lower bound where the distribution should be
truncated.}

\item{upper_bound}{Optional. The upper bound where the distribution should be
truncated.}
}
\description{
Define a prior distributions
}
