% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-priors.R
\name{prior}
\alias{prior}
\title{Define a prior distributions}
\usage{
prior(distribution, type, coefficient = NA, lower_bound = NA, upper_bound = NA)
}
\arguments{
\item{distribution}{A distribution statement for the prior
(e.g., \code{normal(0, 2)}). For a complete list of available distributions, see
the \emph{Stan} documentation at \url{https://mc-stan.org/docs/}.}

\item{upper_bound}{}
}
\description{
Define a prior distributions
}
