% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-class-model-components.R
\name{structural-model}
\alias{structural-model}
\alias{unconstrained}
\alias{independent}
\alias{hdcm}
\title{Structural models for diagnostic classification}
\usage{
unconstrained()

independent()

hdcm()
}
\value{
A structural model object.
}
\description{
Structural models define how the attributes are related to one another.
The currently supported options for structural models are:
unconstrained, independent attributes and HDCM.
See details for additional information on each model.
}
\details{
The unconstrained structural model places no constraints on how the
attributes relate to each other. This is equivalent to a saturated model
described by Hu & Templin (2020) and in Chapter 8 of Rupp et al. (2010).

The independent attributes model assumes that the presence of the attributes
are unrelated to each other. That is, there is no relationship between the
presence of one attribute and the presence of any other. For an example of
independent attributes model, see Lee (2016).

The hierarchical attributes model assumes some attributes must be mastered
before other attributes can be mastered. For an example of the hierarchical
attributes model, see Leighton et al. (2004) and Templin & Bradshaw (2014).
}
\examples{
unconstrained()

independent()

hdcm()
}
\references{
Hu, B., & Templin, J. (2020). Using diagnostic classification
models to validate attribute hierarchies and evaluate model fit in Bayesian
Networks. \emph{Multivariate Behavioral Research, 55}(2), 300-311.
\doi{10.1080/00273171.2019.1632165}

Lee, S. Y. (2016). \emph{Cognitive diagnosis model: DINA model with
independent attributes}.
https://mc-stan.org/documentation/case-studies/dina_independent.html

Rupp, A. A., Templin, J., & Henson, R. A. (2010). \emph{Diagnostic
measurement: Theory, methods, and applications}. Guilford Press.

Leighton, J. P., Gierl, M. J., & Hunka, S. M. (2004). The
attribute hierarchy method for cognitive assessment: A variation on
Tatsuoka's rule-space approach.
\emph{Journal of Educational Measurement, 41}(3), 205-237.
\doi{10.1111/j.1745-3984.2004.tb01163.x}

Templin, J. L., & Bradshaw, L. (2014). Hierarchical diagnostic
classification models: A family of models for estimating and testing
attribute hierarchies. \emph{Psychometrika, 79}(2), 317-339
\doi{10.1007/s11336-013-9362-0}
}
\seealso{
\link[=measurement-model]{Measurement models}
}
