% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-class-model-components.R
\name{structural-model}
\alias{structural-model}
\alias{unconstrained}
\alias{independent}
\title{Structural models for diagnostic classification}
\usage{
unconstrained()

independent()
}
\value{
A structural model object.
}
\description{
Structural models define how the attributes are related to one another.
The currently supported options for structural models are:
unconstrained and independent attributes.
See details for additional information on each model.
}
\details{
Additional details...
}
\examples{
unconstrained()

independent()
}
