% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-class-model-components.R
\name{structural-model}
\alias{structural-model}
\alias{unconstrained}
\alias{independent}
\alias{loglinear}
\title{Structural models for diagnostic classification}
\usage{
unconstrained()

independent()

loglinear(loglinear_interaction = Inf)
}
\arguments{
\item{loglinear_interaction}{For the Log-linear structural model, the highest
structural-level interaction to include in the model.}
}
\value{
A structural model object.
}
\description{
Structural models define how the attributes are related to one another.
The currently supported options for structural models are:
unconstrained; independent attributes; and loglinear.
See details for additional information on each model.
}
\details{
The unconstrained structural model places no constraints on how the
attributes relate to each other. This is equivalent to a saturated model
described by Hu & Templin (2020) and in Chapter 8 of Rupp et al. (2010).

The independent attributes model assumes that the presence of the attributes
are unrelated to each other. That is, there is no relationship between the
presence of one attribute and the presence of any other. For an example of
independent attributes model, see Lee (2016).

The loglinear structural model assumes that class membership proportions can
be predicted using a loglinear model that includes main and interaction
effects (see Xu & von Davier, 2008). A saturated loglinear structural model
includes interaction effects for all attributes measured in the model, and is
equivalent to the unconstrained structural model and the saturated model
described by Hu & Templin (2020) and in Chapter 8 of Rupp et al. (2010). A
reduced form of the loglinear structural model containing only main effects
is equivalent to an independent attributes model (e.g. Lee, 2016).
}
\examples{
unconstrained()

independent()

loglinear()

loglinear(loglinear_interaction = 2)
}
\references{
Hu, B., & Templin, J. (2020). Using diagnostic classification
models to validate attribute hierarchies and evaluate model fit in Bayesian
Networks. \emph{Multivariate Behavioral Research, 55}(2), 300-311.
\doi{10.1080/00273171.2019.1632165}

Lee, S. Y. (2016). \emph{Cognitive diagnosis model: DINA model with
independent attributes}.
https://mc-stan.org/documentation/case-studies/dina_independent.html

Rupp, A. A., Templin, J., & Henson, R. A. (2010). \emph{Diagnostic
measurement: Theory, methods, and applications}. Guilford Press.

Xu, X., & von Davier, M. (2008). \emph{Fitting the structured general
diagnostic model to NAEP data} (RR-08-27). Princeton, NJ: Educational
Testing Service.
}
\seealso{
\link[=measurement-model]{Measurement models}
}
